Add transmute/utils.rs
